Error messages for missing includes and some more tests
