+++ tests/buildErrors/importFileWithError.manifold
@@ -1,0 +1,1 @@
+import "parseError.manifold";
