+++ tests/buildErrors/importFileWithError.manifold.error
@@ -1,0 +1,1 @@
+Error at line 2, char 0: no viable alternative at input 'compile.'
