+++ tests/buildErrors/unknownImport.manifold
@@ -1,0 +1,1 @@
+import "../hello.manifold";
