+++ tests/buildErrors/unknownImport.manifold.error
@@ -1,0 +1,1 @@
+Import ../hello.manifold not found
